requirements.txt
anyio<=4.0.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
exceptiongroup==1.1.3
Flask<=3.0.0
h11==0.14.0
httpcore<=0.18.0
httpx<=0.25.0
idna==3.4
importlib-metadata<=6.8.0
itsdangerous<=2.1.2
gunicorn<=21.2.0
Jinja2<=3.1.2
MarkupSafe==2.1.3
python-dotenv<=1.0.0
python-telegram-bot<=20.6
requests<=2.31.0
sniffio<=1.3.0
urllib3<=2.0.7
Werkzeug<=3.0.1
zipp<=3.17.0